+++ NEWS.md
@@ -1,4 +1,4 @@
-# BEDMatrix 2.0.2.9000
+# BEDMatrix 2.0.3
 
 - Replace testthat with tinytest.
 - Do not mix up minor/major and first/second alleles in documentation.